refactor: mcp、workspace config
